refactor(sendMetrics): add name label only if hostname is not default
to avoid creating new metrics in prometheus and to avoid the extra definition
in the Config.h send the name label only if the hostname differs from the default.

While at it change the passing of the strings to call by reference to
avoid their duplication at every function invocation.

Signed-off-by: Peter Lieven <pl@dlhnet.de>