feat: add http command endpoint (v2)
currently inverter commands are only accessible via MQTT.

This adds a http endpoint /command to the webserver to
access the inverter commands.

You may invoke /command/list for a list of commands.
All commands must be executed as HTTP POST with JSON encoded
parameters in the body. Additionally the Content-Type of the
body must be application/json to avoid CORS attacks.

As this feature is a potential security issue it has to
be explicitly enabled in the config.

Signed-off-by: Peter Lieven <pl@dlhnet.de>